# hr_website/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HR Website API",
    "version": "1.0.0",
    "description": "API documentation for the HR Website backend."
  },
  "components": {
    "schemas": {
...
      "RequestEntity": {
        "type": "object",
        "required": [
          "id",
          "branch_id",
          "user_id",
          "request_type",
          "title",
          "content",
          "status",
          "reject_reason",
          "approver_id",
          "created_at",
          "updated_at",
          "requester"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "description": "Chi nhánh liên quan (public.branches.id)."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "Nhân viên gửi đơn (public.users.id)."
          },
          "request_type": {
            "type": "string",
            "description": "Loại đơn: nghỉ phép, đổi ca, điều chỉnh công, ..."
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "reject_reason": {
            "type": "string",
            "nullable": true
          },
          "approver_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Người duyệt (lấy từ phiên đăng nhập)."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "requester": {
            "type": "object",
            "nullable": true,
            "description": "Nhân viên nhúng từ khoá ngoại requests_user_id_fkey; null khi RLS ẩn hồ sơ.",
            "required": [
              "id",
              "full_name"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "full_name": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi.",
            "example": "Trường `question` là bắt buộc và phải là chuỗi không rỗng."
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/requests": {
      "get": {
        "summary": "List staff requests",
        "description": "The approval queue of SCRUM-41: leave, shift-swap and adjustment requests with the\nrequester name embedded from `requests.user_id`. The value `all` disables a filter, which\nis what the dashboard pills send.\n",
        "tags": [
          "Requests"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "all"
              ]
            },
            "description": "Exact status to filter by; \"all\" disables this filter."
          },
          {
            "in": "query",
            "name": "branch_id",
            "schema": {
              "type": "string"
            },
            "description": "Branch UUID to filter by; \"all\" disables this filter."
          }
        ],
        "responses": {
          "200": {
            "description": "Requests matching the optional filters, most recently updated first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "branch_id is neither a UUID nor \"all\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit a request",
        "description": "Files the caller's own đơn (SCRUM-41: nghỉ phép, đổi ca, điều chỉnh công) as `PENDING` for a\nmanager to review. This is the submit half of the ticket — the queue above can only fill if\nsomebody writes a row, and nothing else in this app inserts into `public.requests`.\nThe requester always comes from the session (`requests_insert_own` refuses a row whose\n`user_id` is not the caller), and every role holds `request:create` **except the\norganization's `OWNER`** (SCRUM-63): the owner sits above the approval chain and reviews the\nqueue instead of filing into it.\n",
        "tags": [
          "Requests"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "branch_id",
                  "request_type",
                  "title"
                ],
                "properties": {
                  "branch_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Chi nhánh liên quan (public.branches.id); một chi nhánh bạn không nhìn thấy được trả về 404."
                  },
                  "request_type": {
                    "type": "string",
                    "enum": [
                      "Nghỉ phép",
                      "Đổi ca",
                      "Điều chỉnh công",
                      "Khác"
                    ],
                    "description": "Loại đơn, đúng giá trị của REQUEST_TYPES trong lib/domain/entities/RequestEntity.ts."
                  },
                  "title": {
                    "type": "string",
                    "maxLength": 120,
                    "example": "Xin nghỉ phép ngày 12/10"
                  },
                  "content": {
                    "type": "string",
                    "maxLength": 500,
                    "nullable": true,
                    "description": "Nội dung chi tiết; bỏ trống hoặc null là đơn không có ghi chú thêm."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored request, with its requester embedded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestEntity"
                }
              }
            }
          },
          "400": {
            "description": "The JSON body is invalid, request_type is unknown, or title/content is blank or too long.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The account has no user profile, the role may not file a request (SCRUM-59 — the organization's OWNER never does, SCRUM-63), the account is not assigned to a branch yet, or it names a branch other than its own (SCRUM-63).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The branch does not exist or is not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}